Some more refactoring of models structure
